Refactored init
