+++ src/aws/lambda/index.js
@@ -2,10 +2,12 @@ const { createLambdaFunction } = require('./createLambdaFunction');
 const { createLambdaPermission } = require('./createLambdaPermission');
 const { createLambdaRole } = require('./createLambdaRole');
 const { lambdaExists } = require('./lambdaExists');
+const { initJadeLambdas } = require('./initJadeLambdas');
 
 module.exports = {
   createLambdaFunction,
   createLambdaPermission,
   createLambdaRole,
   lambdaExists,
+  initJadeLambdas,
 };
